+++ docs/news/whats_new.md
@@ -12,7 +12,7 @@ Released on ...
 
 ### Features
 
-- Ability to add templates via the [`cardTemplate`](/api/config/booking-cardtemplate) and [`infoTemplate`](/api/config/booking-infotemplate) properties
+- Ability to embed templates via the [`cardTemplate`](/api/config/booking-cardtemplate) and [`infoTemplate`](/api/config/booking-infotemplate) properties
 - [`renderType`](/api/config/booking-rendertype) added to handle big data loading
 
 ## Version 1.0.1